// repairhubcoreui/src/app/shared/models/Stores.ts
import { Appointments } from "./Appointments";
import { Centers } from "./Centers";
import { Customers } from "./Customers";
import { DeviceBrands } from "./DeviceBrands";
import { Devices } from "./Devices";
import { Employees } from "./Employees";
import { InventoryMovements } from "./InventoryMovements";
import { Items } from "./Items";
import { ItemTypes } from "./ItemTypes";
import { Orders } from "./Orders";
import { OrdersItem } from "./OrdersItem";
import { PaymentTypes } from "./PaymentTypes";
import { RepairStatus } from "./RepairStatus";
import { SaleItems } from "./SaleItems";
import { Sales } from "./Sales";
import { ServiceOrderRequeste } from "./ServiceOrderRequeste";
import { ServiceOrders } from "./ServiceOrders";
import { ServiceTypes } from "./ServiceTypes";
import { SODiagnostic } from "./SODiagnostic";
import { SOItems } from "./SOItems";
import { SONotes } from "./SONotes";

export interface Stores {
  id: number;
  centerId: number;
  storeCode: string;
  storeName: string;
  country: string;
  address: string;
  zipCode: string;
  city: string;
  state: string;
  time_zone: string;
  email: string;
  phoneNumber: string;
  webSite: string;
  logo: string;
  createdAt: Date;
  appointments?: Appointments[];
  customers?: Customers[];
  deviceBrands?: DeviceBrands[];
  devices?: Devices[];
  employees?: Employees[];
  inventoryMovements?: InventoryMovements[];
  itemTypes?: ItemTypes[];
  items?: Items[];
  orders?: Orders[];
  ordersItems: OrdersItem[];
  paymentTypes?: PaymentTypes[];
  repairStatuses?: RepairStatus[];
  sODiagnostics?: SODiagnostic[];
  sOItems?: SOItems[];
  sONotes?: SONotes[];
  saleItems?: SaleItems[];
  sales?: Sales[];
  serviceOrderRequestes?: ServiceOrderRequeste[];
  serviceOrders?: ServiceOrders[];
  serviceTypes?: ServiceTypes[];
  center?: Centers;
}